requirements.txt
alpyca==2.0.4
anyio==4.0.0
appnope==0.1.3
astropy==5.3.3
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.5
certifi==2023.7.22
charset-normalizer==3.2.0
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
enum-tools==0.9.0.post1
exceptiongroup==1.1.3
executing==1.2.0
fonttools==4.43.0
future==0.18.3
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
imagecodecs==2023.1.23
imageio==2.31.5
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.16.0
iso8601==2.1.0
jedi==0.19.0
Jinja2==3.1.2
jplephem==2.19
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.5
Logbook==1.6.0
looseversion==1.3.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
munkres==1.1.4
nest-asyncio==1.5.6
netifaces==0.11.0
numpy==1.26.0
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
PIMS==0.6.1
pip==23.2.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyerfa==2.0.0.3
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
Represent==1.6.0.post0
requests==2.31.0
rush==2021.4.0
scipy==1.11.3
serial==0.0.97
setuptools==68.0.0
sgp4==2.22
six==1.16.0
skyfield==1.46
slicerator==1.1.0
sniffio==1.3.0
spacetrack==1.0.1
stack-data==0.6.2
tifffile==2023.9.26
tornado==6.3.3
tqdm==4.66.1
trackpy==0.6.1
traitlets==5.10.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
wcwidth==0.2.7
wheel==0.38.4
zipp==3.17.0
